+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - Support GDK 0.0.46
+- Support for creating and restoring Singlesig wallets on Liquid & Liquid Testnet
 - Sweep paper wallet UI
 - New Camera QR Scanner
 - Handle BIP-21 payment uris from the Android OS (bitcoin: & liquidnetwork:)